# docs/swagger.yaml
basePath: /api/v1
definitions:
  Gommunity_platform_community_interfaces_rest_resources.ErrorResponse:
    properties:
      error:
        example: Invalid request
        type: string
    type: object
...
  resources.CommunityResource:
    properties:
      bannerUrl:
        example: https://example.com/banner.jpg
        type: string
      communityId:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      createdAt:
        example: "2025-11-13T17:02:46Z"
        type: string
      description:
        example: A community for data science enthusiasts to share knowledge and collaborate
        type: string
      iconUrl:
        example: https://example.com/icon.jpg
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      isPrivate:
        example: false
        type: boolean
      name:
        example: Data Science Community
        type: string
      ownerId:
        example: 550e8400-e29b-41d4-a716-446655440002
        type: string
      updatedAt:
        example: "2025-11-13T17:02:46Z"
        type: string
    type: object
  resources.CreateCommunityResource:
    properties:
      bannerUrl:
        example: https://example.com/banner.jpg
        type: string
      description:
        example: A community for data science enthusiasts to share knowledge and collaborate
        maxLength: 500
        minLength: 10
        type: string
      iconUrl:
        example: https://example.com/icon.jpg
        type: string
      isPrivate:
        example: false
        type: boolean
      name:
        example: Data Science Community
        maxLength: 100
        minLength: 3
        type: string
    required:
    - description
    - name
    type: object
...
  resources.UpdateCommunityPrivacyResource:
    properties:
      isPrivate:
        example: true
        type: boolean
    required:
    - isPrivate
    type: object
  resources.UpdateCommunityResource:
    properties:
      bannerUrl:
        example: https://example.com/banner.jpg
        type: string
      description:
        example: Updated community description
        maxLength: 500
        minLength: 10
        type: string
      iconUrl:
        example: https://example.com/icon.jpg
        type: string
      name:
        example: Updated Community Name
        maxLength: 100
        minLength: 3
        type: string
    type: object
...
host: localhost
info:
  contact: {}
  description: Community management API with Kafka event processing
  title: Gommunity API
  version: "1.0"
paths:
  /api/v1/communities:
    get:
      consumes:
      - application/json
      description: Get all communities (paginated)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/resources.CommunityResource'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Gommunity_platform_community_interfaces_rest_resources.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Gommunity_platform_community_interfaces_rest_resources.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all communities
      tags:
      - communities
    post:
      consumes:
      - application/json
      description: Create a new community (only for ROLE_TEACHER and ROLE_ADMIN)
      parameters:
      - description: Community creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/resources.CreateCommunityResource'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/resources.CommunityResource'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Gommunity_platform_community_interfaces_rest_resources.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Gommunity_platform_community_interfaces_rest_resources.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/Gommunity_platform_community_interfaces_rest_resources.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Gommunity_platform_community_interfaces_rest_resources.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new community
      tags:
      - communities
...
  /api/v1/communities/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a community (only owner can delete)
      parameters:
      - description: Community ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Gommunity_platform_community_interfaces_rest_resources.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Gommunity_platform_community_interfaces_rest_resources.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/Gommunity_platform_community_interfaces_rest_resources.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Gommunity_platform_community_interfaces_rest_resources.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Gommunity_platform_community_interfaces_rest_resources.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete community
      tags:
      - communities
    get:
      consumes:
      - application/json
      description: Get a specific community by its ID
      parameters:
      - description: Community ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resources.CommunityResource'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Gommunity_platform_community_interfaces_rest_resources.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Gommunity_platform_community_interfaces_rest_resources.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Gommunity_platform_community_interfaces_rest_resources.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Gommunity_platform_community_interfaces_rest_resources.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get community by ID
      tags:
      - communities
    put:
      consumes:
      - application/json
      description: Update community name, description, icon and banner (only owner
        can update)
      parameters:
      - description: Community ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Community update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/resources.UpdateCommunityResource'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resources.CommunityResource'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Gommunity_platform_community_interfaces_rest_resources.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Gommunity_platform_community_interfaces_rest_resources.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/Gommunity_platform_community_interfaces_rest_resources.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Gommunity_platform_community_interfaces_rest_resources.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Gommunity_platform_community_interfaces_rest_resources.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update community information
      tags:
      - communities
  /api/v1/communities/{id}/privacy:
    patch:
      consumes:
      - application/json
      description: Update the privacy status of a community (only owner can update)
      parameters:
      - description: Community ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Community privacy update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/resources.UpdateCommunityPrivacyResource'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resources.CommunityResource'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Gommunity_platform_community_interfaces_rest_resources.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Gommunity_platform_community_interfaces_rest_resources.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/Gommunity_platform_community_interfaces_rest_resources.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Gommunity_platform_community_interfaces_rest_resources.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Gommunity_platform_community_interfaces_rest_resources.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update community privacy status
      tags:
      - communities
  /api/v1/communities/my-communities:
    get:
      consumes:
      - application/json
      description: Get all communities owned by the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/resources.CommunityResource'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Gommunity_platform_community_interfaces_rest_resources.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Gommunity_platform_community_interfaces_rest_resources.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get my communities as owner
      tags:
      - communities
...
securityDefinitions:
  BearerAuth:
    description: Enter your JWT token directly (Bearer prefix is optional).
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"